add login fragment fixme: implementation loginfragment.java not ok
