+++ app/build.gradle
@@ -36,4 +36,5 @@ dependencies {
     implementation 'com.google.code.gson:gson:2.8.5'
     implementation 'com.squareup.okhttp3:okhttp:3.11.0'
     implementation 'com.android.support:preference-v14:26.1.0'
+    implementation 'com.google.android.gms:play-services-auth:15.0.1'
 }
