+++ app/src/main/java/xyz/eeckhout/smartcity/controller/AccountFragment.java
@@ -1,0 +1,35 @@
+package xyz.eeckhout.smartcity.controller;
+
+import android.arch.lifecycle.ViewModelProviders;
+import android.os.Bundle;
+import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
+import android.support.v4.app.Fragment;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import xyz.eeckhout.smartcity.R;
+
+public class AccountFragment extends Fragment {
+
+    private AccountViewModel mViewModel;
+
+    public static AccountFragment newInstance() {
+        return new AccountFragment();
+    }
+
+    @Override
+    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container,
+                             @Nullable Bundle savedInstanceState) {
+        return inflater.inflate(R.layout.account_fragment, container, false);
+    }
+
+    @Override
+    public void onActivityCreated(@Nullable Bundle savedInstanceState) {
+        super.onActivityCreated(savedInstanceState);
+        mViewModel = ViewModelProviders.of(this).get(AccountViewModel.class);
+        // TODO: Use the ViewModel
+    }
+
+}
