+++ app/src/main/java/xyz/eeckhout/smartcity/controller/LoginViewModel.java
@@ -1,0 +1,7 @@
+package xyz.eeckhout.smartcity.controller;
+
+import android.arch.lifecycle.ViewModel;
+
+public class LoginViewModel extends ViewModel {
+    // TODO: Implement the ViewModel
+}
