+++ app/src/main/java/xyz/eeckhout/smartcity/controller/MapsFragment.java
@@ -184,7 +184,7 @@ public class MapsFragment extends Fragment implements OnMapReadyCallback {
                 @Override
                 public boolean onMarkerClick(final Marker marker) {
                     // Check if a click count was set, then display the click count.
-                    Toast.makeText(getContext(), marker.getTitle() + " has been clicked :" + marker.getTag() + ".", Toast.LENGTH_LONG).show();
+                    Toast.makeText(getContext(), marker.getTitle() + " has been clicked.", Toast.LENGTH_LONG).show();
 
                     // Return false to indicate that we have not consumed the event and that we wish
                     // for the default behavior to occur (which is for the camera to move such that the
@@ -195,7 +195,7 @@ public class MapsFragment extends Fragment implements OnMapReadyCallback {
             map.setOnPolylineClickListener(new GoogleMap.OnPolylineClickListener() {
                 @Override
                 public void onPolylineClick(Polyline polyline) {
-                    Toast.makeText(getContext(), polyline.getTag().toString(), Toast.LENGTH_LONG).show();
+                    Toast.makeText(getContext(), ((xyz.eeckhout.smartcity.model.villeNamur.bikeRoute.Record) polyline.getTag()).getFields().getNom(), Toast.LENGTH_LONG).show();
                 }
             });
         }
