+++ app/src/main/res/layout/fragment_map_setting.xml
@@ -4,6 +4,6 @@
     android:id="@+id/frame_settings_fragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Map_settingFragment">
+    tools:context=".controller.Map_settingFragment">
 
 </FrameLayout>
