+++ app/src/main/res/menu/activity_main_drawer.xml
@@ -7,7 +7,7 @@
         <item
             android:id="@+id/nav_camera"
             android:icon="@drawable/ic_menu_camera"
-            android:title="Carte" />
+            android:title="@string/nav_label_map" />
         <item
             android:id="@+id/nav_gallery"
             android:icon="@drawable/ic_menu_gallery"
@@ -19,7 +19,7 @@
         <item
             android:id="@+id/nav_manage"
             android:icon="@drawable/ic_menu_manage"
-            android:title="Mon Compte" />
+            android:title="@string/nav_label_account" />
     </group>
 
     <item android:title="Communicate">
