+++ app/src/main/res/values/strings.xml
@@ -1,13 +1,14 @@
 <resources>
     <string name="app_name">smartcity</string>
-
+    <!--Navigation-->
     <string name="navigation_drawer_open">Open navigation drawer</string>
     <string name="navigation_drawer_close">Close navigation drawer</string>
     <string name="nav_header_title">Android Studio</string>
     <string name="nav_header_subtitle">android.studio@android.com</string>
     <string name="nav_header_desc">Navigation header</string>
     <string name="action_settings">Settings</string>
-
+    <string name="nav_label_map">Carte</string>
+    <string name="nav_label_account">Mon compte</string>
     <string name="title_activity_maps">Map</string>
     <string name="title_activity_login">Sign in</string>
     <!-- Strings related to login -->
@@ -22,7 +23,7 @@
     <string name="permission_rationale">"Contacts permissions are needed for providing email
         completions."
     </string>
-    <string name="error_internet_connection">Internet connection required !</string>
+    <!--Setting-->
     <string name="map_setting_filter">Filtre</string>
     <string name="map_setting_filter_description">Filtrer les points à afficher sur la carte</string>
     <string name="map_setting_key_isJCDecauxLoadingEnable">isJCDecauxLoadingEnable</string>
@@ -33,5 +34,14 @@
     <string name="map_setting_BikeParking_label">Parking vélos</string>
     <string name="map_setting_BikeRoute_label">Itinéraires vélo</string>
     <string name="map_setting_jcdecaux_label">Libiavélo</string>
+    <!--Error page-->
+    <string name="error_internet_connection">Internet connection required !</string>
     <string name="error_refresh_button_label">Rafraichir</string>
+
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
+    <string name="login_email_label">Email</string>
+    <string name="login_password_label">Mot de passe</string>
+    <string name="login_button_label">Se connecter</string>
+    <string name="login_description_label">Connectez-vous à l\'application afin d\'utiliser toutes les fonctionnalités offertes</string>
 </resources>
